challenger_watcher/pyproject.toml
[project]
name = "challenger-watcher"
version = "0.1.0"
description = "Autonomous challenger bot for API integrity verification"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "web3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff",
    "black",
]

[tool.ruff]
line-length = 100

[tool.ruff.config]
extend-ignore = ["E501"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
